Restore LF for aria.html and prevent future accidental changes
When the aria-kbdshortcuts property was merged into master, the
line endings were accidentally changed from LF to CRLF for the
entire aria.html file. This commit undoes that line-ending change
and adds a .gitattributes file to prevent it from happening in the
future. Note that I am deliberately doing this only for aria.html
because I am not an editor of other specs in this repo. Editors
who wish to address the same issue should add their file(s) to
the existing .gitattributes file. For more information, see:
https://help.github.com/articles/dealing-with-line-endings